# apps/frontend/src/assets/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "企业级 Go+Gin+Gorm Web 后端项目模板",
    "title": "Gin Admin Starter API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "分页获取用户列表",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["用户管理"],
        "summary": "获取用户列表",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 10,
            "description": "每页数量",
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/definitions/go-admin-starter_internal_model.User"
                          }
                        },
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "更新用户信息",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["用户管理"],
        "summary": "更新用户信息",
        "parameters": [
          {
            "description": "用户信息",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/go-admin-starter_internal_model.User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/go-admin-starter_internal_model.User"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    },
    "/api/v1/users/{id}": {
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "删除指定用户",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["用户管理"],
        "summary": "删除用户",
        "parameters": [
          {
            "type": "integer",
            "description": "用户ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "go-admin-starter_internal_model.User": {
      "type": "object",
      "required": ["username"],
      "properties": {
        "createdAt": {
          "type": "string"
        },
        "createdBy": {
          "type": "integer"
        },
        "deletedAt": {
          "$ref": "#/definitions/gorm.DeletedAt"
        },
        "email": {
          "description": "邮箱，唯一索引",
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "nickname": {
          "description": "昵称",
          "type": "string"
        },
        "phone": {
          "description": "手机号",
          "type": "string"
        },
        "status": {
          "description": "状态，1: 正常, 2: 禁用",
          "type": "integer"
        },
        "updatedAt": {
          "type": "string"
        },
        "updatedBy": {
          "type": "integer"
        },
        "username": {
          "description": "用户名，唯一索引",
          "type": "string"
        }
      }
    },
...
    "go-admin-starter_pkg_response.Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {},
        "message": {
          "type": "string"
        }
      }
    },
    "gorm.DeletedAt": {
      "type": "object",
      "properties": {
        "time": {
          "type": "string"
        },
        "valid": {
          "description": "Valid is true if Time is not NULL",
          "type": "boolean"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "请输入 JWT Token，格式: Bearer {token}",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}